Bug #22270139: SERVER CRASH WHEN TRYING TO COUNT THE                RESULTS OF A SUBQUERY USING FULLTEXT
Post-push fix.

A valgrind build is required to repro this issue.


Issue:
-----
In a PREPARE statement, the "hints" object of every match
function is initialized to a dummy value during the
fix_fields phase. The hints object is not always deleted in
the cleanup phase of the match function. It is freed along
with the rest of the memroot at the end of the prepare.

In the EXECUTE statement, the hints of the match function
in the WHERE clause is re-initialized. The hints of select
list's match function are not initialized since derived table
has now been merged into the outer query block.

So the select list's match function has a dangling pointer
for hints.

SOLUTION:
---------
Currently hints are created on the runtime memroot. It
should be created on the prepared statement memroot. This
way it will be deleted only when the prepared statement is
deallocated.