BUG#22469459 WRONG VALUES IN ADDED INDEX WHILE DROPPING VIRTUAL COLUMN
It's a regression of BUG#22082762 RE-ENABLE SUPPORT FOR ADDING VIRTUAL
INDEX WHILE DROPPING VIRTUAL COLUMN

The root cause is that we use altered mysql table object to evaluate
virtual column, the dict table is unchanged before commit. So the
optimizer callback function cannot evaluate the virtual column value.

The solutions is using old mysql table object for the evaluation.

Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>
RB: 11408