Fixed bug#22195458: GCOLS: ASSERTION 0 AND CORRUPTION...
There is a different setting on warning requirment when doing INSERT and
SELECT.  When doing INSERT, server askes to throw warning. But for SELECT, it
askes to ignore warning. Because of the inconsistence, the value of generated
column can be evaluted into different value(e.g YEAR('a')).

In order to keep the consistence, this patch makes YEAR value not to be
related with the warning is required or not.

There is a side effect after this patch. It will make YEAR('a') and
YEAR('aaaa') return the same value '0000'.