Bug#22343301: Error 1093 (HY000): You can't specify target table'.' for update
In the problem query, a derived table is contained in the SET clause of
the UPDATE statement. According to documentation, such derived tables
should be materialized, in order to avoid error about updating a table
that is also read in the same query.

These two mechanisms are used to ensure proper handling:

1. When resolving derived tables, exclude_from_table_unique_test is set
   so that materialized derived tables are excluded from later
   uniqueness test.
2. unique_table() later performs the uniqueness test.

After refactoring of derived tables in 5.7, exclude_from_table_unique_test
is set later, when setup_fields() is called for update_value_list.
But this is too late for unique_table() to notice it.

The solution is to delay the call to unique_table() to after the
derived tables are properly resolved. For single-table UPDATE, this
means after setup_fields() has been called for update_value_list.
For multi-table UPDATE, it means moving the test to
Query_result_update::prepare(), just after setup_fields() has been
called for *values.
Notice also that since Query_result_update::prepare() is not called
during query preparation, the uniqueness test has to be done twice.

Some light refactoring has been made to Query_result_update::prepare(),
to enhance locality of scope of local variables.