Bug#13915291 NOT IN (SUBQUERY) GIVES WRONG RESULT WHEN PREFIX INDEX OF INNER TABLE IS USED
When we switch to index_subquery strategy, remove_additional_cond()
removes equalities injected by in-to-exists, saying that they are
already implemented by the index ref access. Without realizing that
this is wrong reasoning when the index is only on a prefix of the
column.
In the bug's testcase remove_additional_cond() should leave the
conditions in place, instead of removing them and giving a wrong
result (the prefix index gives false positives, causing NOT IN to give
false negatives).

index_subquery (and unique_subquery) is for two types of subqueries:
- subqueries which have to deal with inner NULLs
- subqueries which don't.
The former have more complex injected conditions, use ref_or_null, and
are submitted to remove_additional_cond(). The latter are simpler, use
[eq_]ref, and are submitted to remove_subq_pushed_predicates().
remove_subq_pushed_predicates() properly detects the index prefix
problem, using test_if_ref().
Adding the same logic to remove_additional_cond() proved to be
difficult:
- one has to scan the WHERE, analyze AND-ed parts, a part can be a
triggered condition, itself wrapping two OR-ed conditions
("outer=inner OR inner IS NULL")
- after locating the "outer=inner" equality, we need to see if this
equality is handled by the index; but test_if_ref() will say no for
two reasons: ref_or_null is used and the key has guarded conditions
(see part_of_refkey()). So a modification of test_if_ref() would be
needed; but this function is sensitive (see its comments), and a
modification of part_of_refkey() would be needed too.
- And we cannot just skip test_if_ref() because it handles: prefix
index detection, and truncation detection (example: lookup "aa" in a
CHAR(1) indexed column: "a" will be a match; test_if_ref() notices
that). Those two detections are needed to decide if we can skip the
condition.

For simplification, in this patch remove_additional_cond() is simply
deleted. The drawback is that we have more "using WHERE" for
complete-index (non-prefix) index_subquery using ref_or_null. I think
that it isn't a real problem; indeed, in this equivalent join query:
explain format=json select * from t1 force index (xx) where a1='b' or a1 is null;
shows:
|  1 | SIMPLE      | t1    | NULL       | ref_or_null | xx            | xx   | 9       | const |    2 |   100.00 | Using index condition |
the conditions are still tested, they are not removed.
The HAVING clause isn't affected by this patch.

in_additional_cond is thus unnecessary.
Removed a useless comment from sql_optimizer and a wrong one from
item_subselect.

After removing in_additional_cond, we also remove in_having_cond.
The logic is: in_having_cond is only tested in sql_optimizer.cc,
to see if HAVING is equal to the one created by IN->EXISTS (which may
be false, HAVING can be <other cond> AND <the one created by IN->EXISTS>).
Instead of testing that, we can test if HAVING was created by
IN->EXISTS. That can be done with created_by_in2exists().
A few calls to set_created_by_in2exists() were missing, they're added.
Because this function belongs to Item_bool_func, and objects
like having_item are Item*, their type is change to Item_bool_func*;
it's not a problem as in fact they are Item_bool_func* by the way
they are constructed (AND chains of equalities, triggered conditions and
IS NULL tests). They are constructed by and_items(), so an overload
of and_items() is added, taking Item_bool_func and returning
Item_bool_func.