Bug #22262706: SERVER CRASH IF MTR RUNS FOR EMBEDDED SERVER   AND SPECIFIED UNKNOWN OPTION
After checking the options the server is calling unireg_abort()
and expects that the program will end at that point.
But for the embedded server it was not calling exit() (as the
non embedded branch was doing.
Hence the code after the check for invalid options was invoked with
wrong or missing arguments.
Fixed by adding a call to the exit() function.
No test case is possible due to a restriction in mysqltest
that in embedded mode is unable of starting another program that
is also linked with the embedded server code.