Fixed bug#22195364: GCOLS: FAILING ASSERTION: DFIELD_IS_NULL(DFIELD2) || DFIELD2->DATA
If one column(generated column or non-generated column) with BLOB type is
defined as NOT NULL, the materialized value of prefixed index on it is an
empty string when insert an empty string. When doing a SELECT, the not-null
columns will be transformed into an emptry string during send the result to
client. Before send to client, the value of not-null BLOB column is NULL.

When doing an INSERT with UPDATE on duplicated index, the server will scan the
index first and see whether there is a duplication. If it has, server will do
an UPDATE. For the issued query in the bug report, server tries to use the
index on the not-null BLOB virtual generated column with an empty string
expression to locate the index entry. However, the value of it is a NULL
during evaluate the generated column. When use the NULL value to locate the
index entry, the assertion is triggered because the materizlied prefixed index
value is an empty string.

The above inconsistency is because there is one optimization in
Field_blob::store, which sets the memory to be zero when the length is 0.  In
this patch, we remove such an optimization.