Bug#17270896: Crash in Item_ident::fix_after_pullout()
The crash occurs because an Item_aggregate_ref object has
context.select_lex == depended_from. However, by convention when the
query block an Item_ident object is contained in is the same
query block that it is depended upon, depended_from should be NULL.

Item_ident::fix_after_pullout() assumes that these pointers are different
when depended_from is non-NULL, enters the loop to set outer reference
information, and eventually follows an invalid pointer.

The problem lies within Item::split_sum_func2(), where an
Item_aggregate_ref object that wraps the set function is created.
It is created with context.select_lex equal to the aggregation query
block, and depended_from is the same query block, when this query block
is different from the "base" query block of the set function.

The solution is to create the Item_aggregate_ref with context.select_lex
set to be the "base" query block of the set function, and depended_from
as the aggregation query block. (When the referenced expression is not
a set function, we can safely use the current query block and set
depended_from = NULL.)

The change is made simpler by some refactoring. A new field
base_select is added to class Item_sum_func, holding the container
query block of the set function. The field aggr_sel is renamed to
aggr_select, for consistency. The fields nest_level and aggr_level
have been eliminated.

The function Item_sum_func::depended_from() could also be eliminated.

class Item_aggregate_ref now gets value of depended_from from constructor.

The Item_sum member max_arg_level has been renamed to max_aggr_level.
This is because it keeps track of the maximum level of query blocks for
which aggregation of the set function is possible.

Item_sum::init_sum_func_check() and Item_sum::check_sum_func() have
been refactored:

- Comments have been polished. Reference to extension relative to
  SQL2011 standard has been added.

- Logic for calculating validity of set function in check_sum_func()
  has been simplified: Value of aggr_select is now keeping track of
  whether we have a valid aggregation query, NULL means no valid
  aggregation has been found as yet.
  The tests are more logical: First check aggregation at same level
  as set function is placed in, then check aggregation at inner level,
  then check aggregation of column-less set function.

- register_sum_func() has been deleted and functionality moved to
  check_sum_func(). This was done because the second part of
  register_sum_func() was done prematurely, since check_sum_func()
  may determine an aggregating query after register_sum_func(), and
  in some cases mark_as_dependent() was called with NULL as argument.
  Fixing this caused a change in test subquery, where the middle
  query block was erroneously marked as dependent.