Bug #21807286: "CREATE USER IF NOT EXISTS" REPORTS AN ERROR
CREATE|ALTER IF [NOT] EXISTS ... IDENTIFIED BY ... was failing
if the user existed(not existed for alter) and raw error log was
off due to considerations of wrong binlogging in that case.

Converted this to a warning (regardless of the raw log option)
because this is how it operates without the IDENTIFIED BY option.
Logging IDENTIFIED BY is safe, since it's converted on the fly to
IDENTIFIED WITH (hence is not revealing the password).

Also made sure that the warning for a user is not counted towards
the number of users successfully processed by CREATE USER.

Also made sure that the IDENTIFIED clauses are handled properly
when binlogged.

Added test case and updated the existing tests to cover the change.

Now logging of all user operations involving credentials that
are a no-op on the master is inhibited.