BUG#22523554 - CRASH AFTER FAILURE IN RELAY-LOG-RECOVERY, ON START GR COMMAND
Problem :

The problem here is that by setting the relay log recovery to 1 we are
getting a an error during recovery. Once the relay log recovery is
happening the repositories are not initiated and the server reports
error due to this. However if we try to do a START GROUP_REPLICATION in
this situation, the server crashes as it uses the "mi->rli" to fetch the
current log position and log name.

Fix :

The fix makes sure that the recovery procedure will not take
group_replication_channel into consideration. It will skip the relay
logs of the GR channel and proceed further.