BUG#22305605 STOP SLAVE IO THREAD PRINTS WRONG LOST CONNECTION MESSAGE IN ERROR LOG FILE
Problem:
STOP SLAVE IO_THREAD closes socket communication between Master and Slave.
This prints an ERROR message in the error log
[ERROR] Error reading packet from server: Lost connection to
MySQL server during query (server_errno=2013). Here the socket closed intentionally
by the DBA using command 'STOP SLAVE IO_THREAD'. Hence no need to print the message
[ERROR] that says "Lost connection" which will confusion the users/DBAs.

Analysis & Fix: When we are unable to read packet from the replication channel,
we print this error message under the assumption that we lost connection
from the master. But this is not true always, it could be that replication channel
is closed due to a 'STOP SLAVE IO_THREAD' command execution on Slave.
In this situation, this panic error should not be printed. Now after the patch,
slave server checks if abort_slave flag is set or not. If it is not set,
then it assumes that replication channel lost connection from
the master and print this error message in the error log.