BUG#20836250 MYSQLBINLOG --VERBOSE FOR BINARY DATATYPE NOT SKIPPING ' (QUOTE) & \ (BACKSLASH)
Problem & Analysis:
===================
Character set information is not available in the binary log,
which affects string/binary column display in mysqlbinlog --verbose:
There is no distinction made between corresponding binary and
nonbinary string types (BINARY and CHAR, VARBINARY and VARCHAR, BLOB
and TEXT). The output uses a data type of STRING for
fixed-length strings and VARSTRING for variable-length strings.
But single quote (' which 0x27) and backslash (\ which is 0x5c)
are possible in binary datatype content and if we display single
quote as character it will confuse the users and also backslash will
skip the next character if we are treating them as string value.

Fix: Display hexdecimal equivalent value for single quote and backslash,
dont treat them as normal string characters i.e., display \x27 or \x5c
instead.