Bug #22274985: SEGMENTLIST UNIT TEST CAUSING BUILD ERROR
 - Link failure when compiling the SegmentList-t unit test
   since it didn't link the NdbGetRUsage_Init() and
   NdbGetRUsage_End() functions.
 - When examining the problem it was found that the above mentioned
   two functions were not necessary since the mach_task_self()
   function does not allocate any resources. See
   https://robert.sesek.com/2012/1/debugging_mach_ports.html:

"Earlier in the week I had audited all instances of Mach calls in
the code base, including two to mach_thread_self(). This call is
very similar to mach_task_self(), but instead of returning the task
port, it returns the thread port. On first call, though,
mach_thread_self() will also obtain a send right on the port, which
must be deallocated manually. The mach_task_self() function does not
have this requirement. In terms of API design, this is a little
confusing because two similarly named methods have very different
ownership semantics. The reason for this is an odd implementation
detail:

In xnu/libsyscall/mach/headers/mach_init.h, Apple #defines
mach_task_self() to be the extern’d global mach_task_self_, a
cached copy of the task port. Because it’s cached, the send right
is already held and does not get incremented when you invoke the
macro. If you were to call task_self_trap(), though, a call into
the kernel would occur and a new Mach port with a send right would
be returned. Because mach_thread_self() is not cached in this
manner, you need to release the send right with a call to
mach_port_deallocate()."

 - fix by removing the two new unnecesary functions.