Bug#22097534  SUPER_READ_ONLY ABORTS STOP SLAVE IF RELAY_LOG_INFO_REPOSITORY=TABLE, DBG CRASH
If relay_log_info_repository=TABLE, STOP SLAVE results in
'STOP SLAVE' failed: 1290: The MySQL server is running with
the --super-read-only option, so it can not execute this
statement. The problem is caused due to updating slave info
tables when executing command 'STOP SLAVE'. The problem is
also caused due to updating slave info tables (and updating
slave_worker_info table if MTS is enabled) when executing
command "CHANGE MASTER TO".

To fix the problem, permit replication operations regardless
of the settings of read_only and super_read_only.