Bug#22551595: UPDATE WINDOWSCACHE.CMAKE FOR VS2015
WindowsCache.cmake contains hardcoded results of various CMake
checks we do during building. This is useful as running these
tests on Windows is pretty slow.

This patch updates WindowsCache.cmake now that we have
switched to VS2015 for 5.8. Since we now have timespec on
all platforms, our own implementation which was used <VS2015
is removed. Missing entries in WindowsCache.cmake are also added.