Bug#21974696: WRONG RESULT WHEN GROUPING ON VALUE FROM SUBQUERY               AND USING COUNT DISTINCT
Queries that group the results on subqueries which return BLOB, or a
BLOB-based type such as TEXT or JSON, sometimes fail to find the group
boundaries and return wrong results.

This bug was fixed in MySQL 8.0.4 with the fix for
Bug#26188578 Bug#26164633 Bug#26360114 Bug#26781725 Bug#26848089 Regressions with WL#9236.

Adding a test case that verifies that the bug is fixed.