Bug #22329365 NDB_RESTORE: DISABLE UK CHECKS WHEN RESTORING NO DATA
If ndb_restore is run on a backup which contains a unique index,
duplicate key errors can occur. During data restore, the database
can pass through many inconsistent states before the restore
completes. An inconsistent state may have duplicate values for
a column which has a unique index. This makes ndb_restore error
out with duplicate key errors. If the data restore is preceded by
--disable-indexes and followed by --rebuild-indexes, these
errors are avoided.

Added a check for unique indexes in the backup. This check is done
only for data restore. It does not process tables which have
explicitly been excluded. For each unique index found, a warning is
printed.