Bug #21280456: REMOVE DEPRECATED PORTNUMBER CONFIG VARIABLE IN 7.5
Changed TCP.PortNumber, SHM.PortNumber and SCI.Portnumber
to CI_INTERNAL since they are being used for dynamic ports as well
and added a description for the parameters since the
description can't be left empty for CI_INTERNAL parameters.

Removed the deprecation warning that used to appear when
TCP/SCI/SHM PortNumber was specified. Also, deleted the code that
was overwriting the PortNumber with 0 when PortNumber was
specified in the config. Added a require to ensure
TCP/SHM/SCI PortNumber is not present when the configuration
is loaded from the config file.

Introduced a check in the configuration file parser so that
an error is thrown when CI_INTERNAL parameters are set in
the configuration file.