Bug #22582233 NDB : MAXDISKWRITESPEEDOWNRESTART NOT USED
The fix for
  Bug 20204854 BACKUP FAILING UNDER HEAVY LOAD EXCEPT WHEN
  SNAPSHOTSTART USED

contained a bug which resulted in the wrong disk write
speed configuration parameter being used during System
and Node restarts.

MaxDiskWriteSpeedOtherNodeRestart was used instead of
MaxDiskWriteSpeedOwnRestart.

MaxDiskWriteSpeedOwnRestart       (Min=Max,
                                   Default 200MB/s)
MaxDiskWriteSpeedOtherNodeRestart (Default Min=10MB/s,
                                           Max=50MB/s)

This generally results in local checkpoint (LCP) taking
longer than designed during node and system restarts, which
increased the duration of node and system restarts.

This is fixed so that MaxDiskWriteSpeedOwnRestart is
used while a node is restarting.