Bug #22179317 AGAIN, SIGNAL 11 IN INNODB.INNODB_BUFFER_POOL_RESIZE_DEBUG
Problem:

In the function buf_block_align(), there is a check to wait for the
completion of buffer pool resize.  But after this check, there is nothing
to stop the buffer pool resize to happen.  This resulted in SIGSEGV
fault.

Solution:

The variable buf_pool_resizing, must be protected in some way so that
buf_block_align() can prevent a buffer pool resize from happening during
its execution.  The buf_chunk_map_latch read-write lock can be used
for this purpose.

rb#11449 approved by Marko.