WL#7364: RBR: Enhanced Applier Thread Progress Details
This patch adds functionality to track RBR Applier progress
details through the P_S stages framework. It adds three new
stages which are instrumented to report estimated work and
work completed.

In addition, this patch also sets the database that an applier
is working on correctly as well standardizes the command field
to report COM_QUERY instead of a bogus value.

Finally, it fixes a bug in the THD::set_db instrumentation,
refactors slightly the RBR applier code, adds and fixes tests
for the new functionality.