Bug #22328426: NEW TEST NDB_MGM_PROMPT FAILS WITH VALGRIND
The test ndb_mgm_prompt failed with valgrind. The test used the
script utility and was limited to linux machines with the
above mentioned utility installed.

Removed the ndb_mgm_prompt test and added the test case to
ndb_mgm.test. Changed the behavior such that the change in
prompt is reflected in standard output, similar to the
behavior of the mysql client.