Bug#21974346: GROUPING ON AGGREGATED RESULTS NOT ALWAYS REJECTED
If a subquery references the alias of an aggregated expression from
one of the outer query blocks, it could sometimes be accepted even if
the alias was referenced from within the GROUP BY clause of the query
block where the aggregation happened. Such queries are meaningless and
should be rejected.

The problem lies in Item_ref::fix_fields(), which currently rejects
the reference if it is not inside the HAVING clause of the query block
that contains the reference. There are two problems with this:

1) It is not the position within the query block that contains the
reference that should decide if the reference is valid, but the
position inside the query block where the aggregation happens.

2) Non-outer references to aggregated expressions are allowed other
places than HAVING, for example in ORDER BY. Outer references could be
accepted in those positions too. They should be rejected if they are
in the GROUP BY clause, though, since grouping happens before
aggregation.

This patch makes the following changes:

Item_ref::fix_fields() now raises an error if a reference to an
aggregated expression is in the GROUP BY clause of the query block
where the referenced expression is aggregated. As before, it also
raises an error if the reference is not in the HAVING clause of the
query block it is referenced from. The net effect is that we reject
the previously accepted meaningless queries, but we don't accept any
new queries that were previously rejected.

resolve_ref_in_select_and_group() is reorganized to make it clearer
how it works. It now returns earlier when it knows what the result is
going to be, so that it doesn't have to check the same conditions
multiple times. This part of the patch is just a cleanup, and it is
not essential for the correctness of the queries. It was originally
included in the patch in order to remove some dead code for rejecting
invalid forward references, but the dead code was revived in
bug#22328395.

Change-Id: I83a71931f5edf5e016bda5ba48672245d0eb099b