Bug#20988291 XA+MTS: SQL_THREAD FAILS WITH ER1756 'ERROR WRITING SLAVE WORKER CONFIGURATION'
The reported in the bug error occurs at XA PREPARE handling by the
parallel slave applier. Its reason is quite similar to one of Bug #21618727
and is in that the prepared XA transaction state does not allow to add up
any new statement to the transaction.
This time the statement attempts to update 'mysql.slave_worker_info'.

Notice that the bug does **not** happen in the single-threaded applier because
of System_table_access::open_table() is invoked from Log_event::update_pos()
**after** the XA transaction gets prepared. At the end of XA-prepare the slave
applier always reset the transaction state to NOTR (non-existing transaction),
so the system table updating is permitted.

**Fixed** with shifting commit_positions() invocation to after do_commit(),
which effectively reproduces the STS execution pattern.

**Note**
Remaining crash-safety issues, like a crash in between do_commit() and
flush_info(), are to be covered by a separate BUG20672719.