BUG#13687542 : MTR BEHAVES WRONG WHEN EXECUTING 'ERROR $VAR' IN A LOOP
Description :
=============
MTR's 'error' command can be used when a statement is expected to fail.
It is possible to give the error code either hard-coded or as an mtr variable,
like: --error $variable.

However, if an 'error $variable' statement is inside a loop, it evaluates
to the same value every iteration.

Issue :
=======
'expected_errors' array was not getting updated with the 'error' command
value after the first iteration.

Fix :
=====
Check for 'error' command and update the 'expected_errors' array for each
iteration of the loop.

Reviewed-by: Sayantan Dutta <sayantan.dutta@oracle.com>
RB: 11463