Bug #20960839 NDB WINDOWS REG CREATE KEY LOG CHANGE
While setting up event logging on Windows, ndb_mgmd will try to add a key
to HKEY_LOCAL_MACHINE registry. This operation will fail if the user does
not have access to the registry. In this case, ndb_mgmd
logs an error "Could neither create or open key". This log is not accurate.

Mysqld also uses the same function and also hits the same problem.
 Copied the log from the mysqld code