BUG#22216779: 5.6 CRASHES ON A 5.7 DATADIR
ANALYSIS:
=========
Server startup on 5.6 with the 5.7 data directory results in
server crash. This is basically an attempt to downgrade, which
is unsupported.
The system table 'plugin' uses the InnoDB storage engine in 5.7
whereas it uses MyISAM in 5.6. So during 5.6 server startup,
the 'plugin' table cannot be opened which results in the 'hton'
pointer set to NULL. Access to this pointer results in the
server crash.

FIX:
====
A check has been introduced to verify if the 'hton' pointer contains
a non NULL value. If not, safe server shutdown is ensured.